test2/requirements.txt
numpy==1.16.5
networkx==2.3
matplotlib==3.1.1
-e git+https://github.com/projectmesa/mesa#egg=mesa
mesa_SIR >=0.0.3
